+++ CHANGELOG.md
@@ -1,26 +1,56 @@
 # HydraMPP
 
+## Version 0.0.3
+
+### v0.0.3 New Features
+
+- Utility to check the status of HydraMPP
+  - hydra-status.py
+- SLURM helper
+
+### v0.0.3 Enhancements
+
+- better output
+- host initialization improvement
+
+### v0.0.3 Bug Fixes
+
+- Works with multiple nodes
+
+### v0.0.3 Known issues
+
+- SLURM not yet functional
+
+### v0.0.3 Future plan
+
+- password
+  - encryption
+- compression?
+- error checking
+  - if more CPUs are requested than available
+- suppress worker node stdout/stderr, redirect
+
 ## Version 0.0.2
 
-### New Features
+### v0.0.2 New Features
 
 - sockets (network)
 - Distributed Processing
 - pickle objects
 
-### Enhancements
+### v0.0.2 Enhancements
 
 - cleaner code
 
-### Bug Fixes
+### v0.0.2 Bug Fixes
 
 - multiprocessing sync error
 
-### Known issues
+### v0.0.2 Known issues
 
 - Might not work with more than 2 nodes
 
-### Future plan
+### v0.0.2 Future plan
 
 - password
   - encryption
@@ -31,14 +61,14 @@
 
 ## Version 0.0.1
 
-### New Features
+### v0.0.1 New Features
 
 - local MPP
   - Functions as a quick replacement for Ray (minimal options only)
 
-### Bug Fixes
+### v0.0.1 Bug Fixes
 
-### Future plan
+### v0.0.1 Future plan
 
 - sockets (network)
 - Distributed Processing